Add additional escape key sequences for AZERTY layouts
Using the Escape key followed by Shift and a number key to insert a
foreground mosaic attribute doesn't work on an AZERTY keyboard layout where
the Shift key is always needed to type a number.

The additional escape keys are on the top row of letters situated below the
equivalent number key as those letters were not previously not used.

Red:     1 or A
Green:   2 or Z
Yellow:  3 or E
Blue:    4 or R
Magenta: 5 or T
Cyan:    6 or Y
White:   7 or U
Black:   0 or P